+++ module/ldbc-dsl/src/test/scala/ldbc/dsl/DatabaseConnectionTest.scala
@@ -29,7 +29,7 @@ object DatabaseConnectionTest extends Specification:
   ds.setUser("ldbc")
   ds.setPassword("password")
 
-  private val dataSource = DataSourceIO[IO](ds)
+  private val dataSource = DataSource[IO](ds)
 
   given LogHandler[IO] = LogHandler.consoleLogger
